Integracion de cambios y orden de archivos
