+++ index.html
@@ -1,47 +1,43 @@
 <!DOCTYPE html>
 <html lang="es">
-
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>ACREDICOM-FORMS </title>
     <script src="https://cdn.tailwindcss.com"></script>
-    <link href="https://fonts.googleapis.com/css2?family=Poppins:wght@300;400;500;600;700&display=swap"
-        rel="stylesheet">
-    <script src="https://code.iconify.design/1/1.0.0/iconify.min.js"></script>
+    <link href="https://fonts.googleapis.com/css2?family=Poppins:wght@300;400;500;600;700&display=swap" rel="stylesheet">
     <script src="https://cdn.jsdelivr.net/npm/signature_pad@4.0.0/dist/signature_pad.umd.min.js"></script>
-    <script src="https://cdn.jsdelivr.net/npm/axios/dist/axios.min.js"></script>
-    <link href="https://cdn.jsdelivr.net/npm/sweetalert2@11/dist/sweetalert2.min.css" rel="stylesheet">
-    <script src="https://cdn.jsdelivr.net/npm/sweetalert2@11"></script>
-    <script src="https://unpkg.com/@panzoom/panzoom@9.6.0/dist/panzoom.min.js"></script>
+    <script src="src/js/axios.min.js"></script>
+    <script src="src/js/iconify.min.js"></script>
+    <script src="src/js/sweetalert.min.js"></script>
+    <link rel="stylesheet" href="src/css/sweetalert2.min.css">
 
     <style>
         body {
             font-family: 'Poppins', sans-serif;
-            color: #00457C; /* Cambiado a #00457C */
+            color: #00457C;
         }
     </style>
 </head>
-
-<!-- Start of Selection -->
 <body class="min-h-screen bg-gradient-to-r from-blue-10 to-blue-50 animate-gradient">
     <div id="app" class="container mx-auto px-4">
         <div class="p-6 max-w-md mx-auto">
-            <div class="space-y-6">
-                <div class="flex justify-end">
-                    <img src="./logo.svg" alt="Logo" class="h-6 w-auto">
-                </div>
-                <!-- Header -->
+            <div class="space-y-2">
+                <header class="flex items-center justify-between ">
+                    <img src="src/img/header.jpg" alt="">
+                </header>
                 <div class="space-y-1">
                     <h1 class="text-3xl font-extrabold text-[#00457C]">Formularios</h1> <!-- Cambiado a #00457C -->
                     <p id="form-title" class="text-gray-500">Nombre del formulario</p>
-                    <p class="text-[#00457C] font-extrabold text-right font-semibold" id="form-number">Formulario 1/4</p> <!-- Cambiado a #00457C -->
+                    <p class="text-[#00457C] font-extrabold text-right font-semibold" id="form-number">Formulario 1/4
+                    </p> <!-- Cambiado a #00457C -->
                 </div>
 
                 <!-- User Info Card -->
                 <div class="bg-blue-50 p-4 rounded-lg space-y-2">
                     <div class="flex items-center gap-2">
-                        <svg class="w-5 h-5 text-[#00457C]" fill="none" stroke="currentColor" viewBox="0 0 24 24"> <!-- Cambiado a #00457C -->
+                        <svg class="w-5 h-5 text-[#00457C]" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+                            <!-- Cambiado a #00457C -->
                             <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2"
                                 d="M16 7a4 4 0 11-8 0 4 4 0 018 0zM12 14a7 7 0 00-7 7h14a7 7 0 00-7-7z" />
                         </svg>
@@ -49,14 +45,16 @@
                     </div>
                     <div class="flex items-center gap-4">
                         <div class="flex items-center gap-2">
-                            <svg class="w-5 h-5 text-[#00457C]" fill="none" stroke="currentColor" viewBox="0 0 24 24"> <!-- Cambiado a #00457C -->
+                            <svg class="w-5 h-5 text-[#00457C]" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+                                <!-- Cambiado a #00457C -->
                                 <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2"
                                     d="M9 5H7a2 2 0 00-2 2v12a2 2 0 002 2h10a2 2 0 002-2V7a2 2 0 00-2-2h-2M9 5a2 2 0 002 2h2a2 2 0 002-2M9 5a2 2 0 012-2h2a2 2 0 012 2" />
                             </svg>
                             <span id="user-id" class="text-[#00457C]"></span> <!-- Cambiado a #00457C -->
                         </div>
                         <div class="flex items-center gap-2">
-                            <svg class="w-5 h-5 text-[#00457C]" fill="none" stroke="currentColor" viewBox="0 0 24 24"> <!-- Cambiado a #00457C -->
+                            <svg class="w-5 h-5 text-[#00457C]" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+                                <!-- Cambiado a #00457C -->
                                 <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2"
                                     d="M17.657 16.657L13.414 20.9a1.998 1.998 0 01-2.827 0l-4.244-4.243a8 8 0 1111.314 0z" />
                                 <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2"
@@ -86,7 +84,8 @@
                         </button>
                     </div>
                     <button id="submit-form"
-                        class="w-full bg-[#00457C] text-white py-3 rounded-lg font-medium flex items-center justify-center"> <!-- Cambiado a #00457C -->
+                        class="w-full bg-[#00457C] text-white py-3 rounded-lg font-medium flex items-center justify-center">
+                        <!-- Cambiado a #00457C -->
                         Siguiente
                     </button>
                     <button id="btn-anterior-signature"
@@ -111,7 +110,8 @@
                         </svg>
                     </button>
                     <button id="btn-siguiente"
-                        class="w-full bg-[#00457C] text-white py-3 rounded-lg font-medium flex items-center justify-center"> <!-- Cambiado a #00457C -->
+                        class="w-full bg-[#00457C] text-white py-3 rounded-lg font-medium flex items-center justify-center">
+                        <!-- Cambiado a #00457C -->
                         Siguiente
                         <svg class="w-5 h-5 ml-2" fill="none" stroke="currentColor" viewBox="0 0 24 24">
                             <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2"
@@ -123,6 +123,5 @@
         </div>
     </div>
 </body>
-<script src="./main.js"></script>
-
+<script type="module" src="./src/js/main.js"></script>
 </html>
